+++ docs/source/conf.py
@@ -61,7 +61,7 @@ copyright = '2019, Francesco Farina, Andrea Camisa, Andrea Testa, Ivano Notarnic
 author = 'Francesco Farina, Andrea Camisa, Andrea Testa, Ivano Notarnicola, Giuseppe Notarstefano'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.2'
+release = '0.1.3'
 
 
 # List of patterns, relative to source directory, that match files and